LOGR: Split up different types of logging
Also drop std::localtime, we'll do it in post
Also fix the trailing spaces
