+++ MAIN/src/main.cpp
@@ -1,7 +1,7 @@
 #include "UI/UserInterface.hpp"
 #include "DRVR/Driver.hpp"
 #include "CTRL/Controller.hpp"
-#include "LOGR/Logger.hpp"
+#include "LOGR/Trace.hpp"
 
 #include <chrono>
 #include <iostream>
